requirements.txt
anthropic
pdfplumber
pydantic
python-dotenv
fpdf2
